Consolidate measure-related validations into separate module
With configuration options like semi-additive measures and
measure constraints rolling out our suite of measure-specific
validations is growing to the point where having a dedicated
module makes sense.
