+++ metricflow/model/validations/metrics.py
@@ -15,45 +15,6 @@ from metricflow.model.validations.validator_helpers import (
 )
 
 
-class MetricMeasuresRule(ModelValidationRule):
-    """Checks that the measures referenced in the metrics exist."""
-
-    @staticmethod
-    @validate_safely(whats_being_done="checking all measures referenced by the metric exist")
-    def _validate_metric_measure_references(
-        metric: Metric, valid_measure_names: List[str]
-    ) -> List[ValidationIssueType]:
-        issues: List[ValidationIssueType] = []
-
-        for measure_reference in metric.measure_references:
-            if measure_reference.element_name not in valid_measure_names:
-                issues.append(
-                    ValidationError(
-                        context=MetricContext(
-                            file_context=FileContext.from_metadata(metadata=metric.metadata),
-                            metric=MetricModelReference(metric_name=metric.name),
-                        ),
-                        message=f"Invalid measure {measure_reference.element_name} in metric {metric.name}",
-                    )
-                )
-        return issues
-
-    @staticmethod
-    @validate_safely(whats_being_done="running model validation ensuring metric measures exist")
-    def validate_model(model: UserConfiguredModel) -> List[ValidationIssueType]:  # noqa: D
-        issues: List[ValidationIssueType] = []
-        valid_measure_names = []
-        for data_source in model.data_sources:
-            for measure in data_source.measures:
-                valid_measure_names.append(measure.reference.element_name)
-
-        for metric in model.metrics or []:
-            issues += MetricMeasuresRule._validate_metric_measure_references(
-                metric=metric, valid_measure_names=valid_measure_names
-            )
-        return issues
-
-
 class CumulativeMetricRule(ModelValidationRule):
     """Checks that cumulative sum metrics are configured properly"""
